aica: fix up sampling frequencies a bit
this largely fixes the "earrape" problem during the RTC reset screen, although
it's still not entirely correct.

+++ src/libwashdc/sound/sound.c
@@ -106,7 +106,7 @@ static int snd_cb(const void *input, void *output,
     return 0;
 }
 
-void sound_submit_sample(int16_t sample) {
+void sound_submit_sample(int32_t sample) {
     if (outfile)
         fwrite(&sample, sizeof(sample), 1, outfile);
 }
